config: refactor ccw/ui/terminal_colors.go - large term...
Resolves #259
