+++ ccw/ui/terminal_detection.go
@@ -1,22 +1,9 @@
 package ui
 
 import (
-	"fmt"
 	"os"
-	"strconv"
 	"strings"
 )
-
-// TerminalColorInfo holds detected terminal color information
-type TerminalColorInfo struct {
-	Background        string
-	Foreground        string
-	AccentColor       string
-	SupportsTrueColor bool
-	Colors256         bool
-	ThemeType         string // "light", "dark", "auto"
-	DetectionMethod   string
-}
 
 // DetectTerminalColors attempts to detect terminal color scheme from various sources
 func DetectTerminalColors() *TerminalColorInfo {
@@ -242,175 +229,4 @@ func detectFromCommonTerminals() *TerminalColorInfo {
 	}
 
 	return nil
-}
-
-// isLightColor determines if a color is light or dark
-func isLightColor(colorHex string) bool {
-	// Remove # if present
-	colorHex = strings.TrimPrefix(colorHex, "#")
-
-	// Parse RGB values
-	if len(colorHex) != 6 {
-		return false
-	}
-
-	r, err1 := strconv.ParseInt(colorHex[0:2], 16, 0)
-	g, err2 := strconv.ParseInt(colorHex[2:4], 16, 0)
-	b, err3 := strconv.ParseInt(colorHex[4:6], 16, 0)
-
-	if err1 != nil || err2 != nil || err3 != nil {
-		return false
-	}
-
-	// Calculate perceived brightness using standard formula
-	brightness := (0.299*float64(r) + 0.587*float64(g) + 0.114*float64(b)) / 255.0
-	return brightness > 0.5
-}
-
-// CreateThemeFromTerminal creates a color theme based on terminal detection
-func CreateThemeFromTerminal() ColorTheme {
-	colorInfo := DetectTerminalColors()
-
-	var theme ColorTheme
-
-	// Use detected accent color if available
-	accentColor := colorInfo.AccentColor
-	if accentColor == "" {
-		accentColor = "#0066CC" // fallback
-	}
-
-	switch colorInfo.ThemeType {
-	case "dark":
-		theme = ColorTheme{
-			Name:               fmt.Sprintf("terminal-dark (%s)", colorInfo.DetectionMethod),
-			Primary:            accentColor,
-			Success:            "#50C878", // Lighter green for dark background
-			Error:              "#FF6B6B", // Lighter red for dark background
-			Warning:            "#FFD93D", // Lighter yellow for dark background
-			Info:               accentColor,
-			Subtle:             "#888888", // Light gray for dark background
-			Background:         colorInfo.Background,
-			SelectedBackground: accentColor,
-			SelectedForeground: "#FFFFFF",
-			BorderColor:        accentColor,
-		}
-
-	case "light":
-		theme = ColorTheme{
-			Name:               fmt.Sprintf("terminal-light (%s)", colorInfo.DetectionMethod),
-			Primary:            darkenColor(accentColor, 0.3), // Darker for light background
-			Success:            "#006600",                     // Darker green for light background
-			Error:              "#CC0000",                     // Darker red for light background
-			Warning:            "#B8860B",                     // Darker yellow for light background
-			Info:               darkenColor(accentColor, 0.3),
-			Subtle:             "#555555", // Dark gray for light background
-			Background:         colorInfo.Background,
-			SelectedBackground: darkenColor(accentColor, 0.3),
-			SelectedForeground: "#FFFFFF",
-			BorderColor:        darkenColor(accentColor, 0.3),
-		}
-
-	default: // auto or unknown
-		// Use high contrast theme with detected accent color
-		theme = ColorTheme{
-			Name:               fmt.Sprintf("terminal-auto (%s)", colorInfo.DetectionMethod),
-			Primary:            accentColor,
-			Success:            "#00AA00",
-			Error:              "#CC0000",
-			Warning:            "#FF6600",
-			Info:               accentColor,
-			Subtle:             "#666666",
-			Background:         colorInfo.Background,
-			SelectedBackground: accentColor,
-			SelectedForeground: "#FFFFFF",
-			BorderColor:        accentColor,
-		}
-	}
-
-	return theme
-}
-
-// darkenColor darkens a hex color by a given factor (0.0 to 1.0)
-func darkenColor(colorHex string, factor float64) string {
-	colorHex = strings.TrimPrefix(colorHex, "#")
-
-	if len(colorHex) != 6 {
-		return colorHex // Return original if invalid
-	}
-
-	r, err1 := strconv.ParseInt(colorHex[0:2], 16, 0)
-	g, err2 := strconv.ParseInt(colorHex[2:4], 16, 0)
-	b, err3 := strconv.ParseInt(colorHex[4:6], 16, 0)
-
-	if err1 != nil || err2 != nil || err3 != nil {
-		return colorHex // Return original if parse fails
-	}
-
-	// Darken by reducing each component
-	r = int64(float64(r) * (1.0 - factor))
-	g = int64(float64(g) * (1.0 - factor))
-	b = int64(float64(b) * (1.0 - factor))
-
-	// Ensure values stay in valid range
-	if r < 0 {
-		r = 0
-	}
-	if g < 0 {
-		g = 0
-	}
-	if b < 0 {
-		b = 0
-	}
-	if r > 255 {
-		r = 255
-	}
-	if g > 255 {
-		g = 255
-	}
-	if b > 255 {
-		b = 255
-	}
-
-	return fmt.Sprintf("#%02X%02X%02X", r, g, b)
-}
-
-// ShowTerminalDetectionInfo displays what was detected about the terminal
-func ShowTerminalDetectionInfo() {
-	colorInfo := DetectTerminalColors()
-
-	fmt.Println("🖥️  Terminal Color Detection Results")
-	fmt.Println("==================================")
-	fmt.Printf("Detection method: %s\n", colorInfo.DetectionMethod)
-	fmt.Printf("Theme type: %s\n", colorInfo.ThemeType)
-	fmt.Printf("True color support: %v\n", colorInfo.SupportsTrueColor)
-	fmt.Printf("256 color support: %v\n", colorInfo.Colors256)
-
-	if colorInfo.AccentColor != "" {
-		fmt.Printf("Detected accent color: %s\n", colorInfo.AccentColor)
-	}
-
-	if colorInfo.Background != "" {
-		fmt.Printf("Background color: %s\n", colorInfo.Background)
-	}
-
-	if colorInfo.Foreground != "" {
-		fmt.Printf("Foreground color: %s\n", colorInfo.Foreground)
-	}
-
-	fmt.Println("\nEnvironment variables:")
-	relevantVars := []string{
-		"TERM", "COLORTERM", "TERM_PROGRAM", "ITERM_PROFILE",
-		"VSCODE_THEME_KIND", "THEME", "APPEARANCE", "DARK_MODE",
-	}
-
-	for _, varName := range relevantVars {
-		if value := os.Getenv(varName); value != "" {
-			fmt.Printf("  %s = %s\n", varName, value)
-		}
-	}
-}
-
-// GetOptimalTerminalTheme returns the best theme based on actual terminal detection
-func GetOptimalTerminalTheme() ColorTheme {
-	return CreateThemeFromTerminal()
 }
