+++ ccw/ui/terminal_info.go
@@ -1,0 +1,12 @@
+package ui
+
+// TerminalColorInfo holds detected terminal color information
+type TerminalColorInfo struct {
+	Background        string
+	Foreground        string
+	AccentColor       string
+	SupportsTrueColor bool
+	Colors256         bool
+	ThemeType         string // "light", "dark", "auto"
+	DetectionMethod   string
+}
